@@ -28,7 +28,7 @@
   "homepage": "https://github.com/wimpyprogrammer/react-component-update#readme",
   "devDependencies": {
     "babel-cli": "^6.26.0",
-    "babel-jest": "^22.0.1",
+    "babel-jest": "^23.0.0",
     "babel-plugin-transform-object-rest-spread": "^6.26.0",
     "babel-preset-es2015": "^6.24.1",
     "babel-preset-react": "^6.24.1",
@@ -39,20 +39,20 @@
     "enzyme": "^3.1.0",
     "enzyme-adapter-react-16": "^1.0.1",
     "enzyme-adapter-react-helper": "^1.2.2",
-    "eslint": "^4.12.1",
-    "eslint-config-airbnb": "^16.1.0",
-    "eslint-plugin-import": "^2.3.0",
-    "eslint-plugin-jsx-a11y": "^6.0.2",
-    "eslint-plugin-react": "^7.1.0",
-    "jest": "^22.0.1",
+    "eslint": "^5.0.1",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-plugin-import": "^2.13.0",
+    "eslint-plugin-jsx-a11y": "^6.1.0",
+    "eslint-plugin-react": "^7.11.0",
+    "jest": "^23.5.0",
     "lodash.uniqueid": "^4.0.1",
-    "publish-please": "^2.3.1",
+    "publish-please": "^3.1.0",
     "react": "*",
     "react-dom": "*",
     "react-test-renderer": "*",
     "rimraf": "^2.6.1",
-    "sinon": "^4.1.2",
-    "sinon-chai": "^2.13.0"
+    "sinon": "^6.0.1",
+    "sinon-chai": "^3.0.0"
   },
   "peerDependencies": {
     "react": "*"
